// app/src/app/models/maximo.ts
export interface Maximo {
    wonum: string
    description: string
    status: string
    canal: string
    rol_operacion: string
    modificacion_en_curso: string
    creador: string
    grupo_creador: string
    ownergroup: string
    llamar_tecnico: string
    telefono_tecnico: string
    ambito: string
    propietario: string
    servicioti: string
    cinum: string
    ci: string
    sistema: string
    sistema_nombre: string
    impacto: string
    tipo_implementacion: string
    motivo_cambio: string
    desc_impacto: string
    impacto_negocio: string
    validacion_verificacion: string
    afecta_vip: string
    compomentes_criticos: string
    cliente: string
    tipo_cambio: string
    categoria: string
    indisponibilidad: string
    req_ipl: string
    req_batch: string
    req_pi: string
    schedstart: string
    schedfinish: string
    duracion_estimada: string
    incumple_ventana: string
    duracion_ma: string
    fecha_limite_entrega: string
    leadtime: string
    global: string
    alta_coordinada: string
    req_mitiga: string
    req_rollout: string
    cambios_cmdb: string
    cambios_doc: string
    afecta_replica: string
    fecha_creacion: string
    fecha_reportado: string
    description_longdescription: string
  
  }
  
  
  